api/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1     
mysql-connector-python==8.0.32
mysqlclient==2.2.4   
SQLAlchemy==2.0.28
types-Flask==1.1.6
pillow-avif-plugin==1.4.3
bcrypt>=4.0.0
pillow>=10.0.0
pyopenssl
waitress